codex: rename MCP config types module
