+++ codex-rs/config/src/lib.rs
@@ -4,7 +4,7 @@ mod constraint;
 mod diagnostics;
 mod fingerprint;
 mod mcp_edit;
-mod mcp_server;
+mod mcp_types;
 mod merge;
 mod overrides;
 mod project_root_markers;
@@ -54,12 +54,12 @@ pub use diagnostics::io_error_from_config_error;
 pub use fingerprint::version_for_toml;
 pub use mcp_edit::ConfigEditsBuilder;
 pub use mcp_edit::load_global_mcp_servers;
-pub use mcp_server::AppToolApproval;
-pub use mcp_server::McpServerConfig;
-pub use mcp_server::McpServerDisabledReason;
-pub use mcp_server::McpServerToolConfig;
-pub use mcp_server::McpServerTransportConfig;
-pub use mcp_server::RawMcpServerConfig;
+pub use mcp_types::AppToolApproval;
+pub use mcp_types::McpServerConfig;
+pub use mcp_types::McpServerDisabledReason;
+pub use mcp_types::McpServerToolConfig;
+pub use mcp_types::McpServerTransportConfig;
+pub use mcp_types::RawMcpServerConfig;
 pub use merge::merge_toml_values;
 pub use overrides::build_cli_overrides_layer;
 pub use project_root_markers::default_project_root_markers;
